requirements.txt
python-docx==1.1.2
docxtpl==0.18.0
openai==1.54.0
pinyin==0.4.0
